Media objects & YearComparator
Created the objects and set their values. I had to add a generic constructor in the Media class in order to create other constructors in the subclasses. We are supposed to use both the Comparable and Comparator classes. I've used the comparator. Get back to me if we need to make changes on these. Not sure if  I did my compare methods efficiently, but here they are.