I finished the javadoc
I uploaded a  ZIP of the src folder with the finished javadoc for the classes. I did that in a new branch and wasn't able to see the file I uploaded. So there's two branches called Javadoced and Javadoc-ed.  I uploaded the files here to be on the safe side.